+++ nequip/scripts/evaluate.py
@@ -90,7 +90,7 @@ def main(args=None, running_as_script: bool = True):
     )
     parser.add_argument(
         "--output-fields",
-        help="Extra fields to write to the `--output`.",
+        help="Extra fields (names comma separated with no spaces) to write to the `--output`.",
         type=str,
         default="",
     )
@@ -146,9 +146,11 @@ def main(args=None, running_as_script: bool = True):
     if args.output is not None:
         if args.output.suffix != ".xyz":
             raise ValueError("Only .xyz format for `--output` is supported.")
-        args.output_fields = args.output_fields.split(",")
-        assert len(args.output_fields) > 0
+        args.output_fields = [e for e in args.output_fields.split(",") if e != ""]
         output_type = "xyz"
+    else:
+        assert args.output_fields == ""
+        args.output_fields = []
 
     if args.device is None:
         device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
@@ -204,7 +206,7 @@ def main(args=None, running_as_script: bool = True):
         str(args.dataset_config), defaults={"r_max": model_r_max}
     )
     if dataset_config["r_max"] != model_r_max:
-        logger.warn(
+        raise RuntimeError(
             f"Dataset config has r_max={dataset_config['r_max']}, but model has r_max={model_r_max}!"
         )